+++ tests/TraversingLattice1.hs
@@ -10,7 +10,7 @@ meanRepairTime = 0.5
 
 specs = Specs { spcStartTime = 0.0,
                 spcStopTime = 1000.0,
-                spcDT = 100.0,
+                spcDT = 0.1,
                 spcMethod = RungeKutta4,
                 spcGeneratorType = SimpleGenerator }
         
@@ -28,7 +28,7 @@ model =
                    putStrLn ""
                    
      runEventInStartTime $
-       enqueueEventWithIntegTimes
+       enqueueEventWithLatticeTimes
        showLatticeNode
 
      runEventInStopTime $
@@ -36,6 +36,6 @@ model =
 
 main :: IO ()
 main =
-  do lattice <- newRandomLattice 5
-     runLIO lattice $
+  do lat <- newRandomLattice 4
+     runLIO lat $
        runSimulation model specs