+++ tests/TraversingLattice3.hs
@@ -10,7 +10,7 @@ meanRepairTime = 0.5
 
 specs = Specs { spcStartTime = 0.0,
                 spcStopTime = 1000.0,
-                spcDT = 400.0,
+                spcDT = 0.1,
                 spcMethod = RungeKutta4,
                 spcGeneratorType = SimpleGenerator }
         
@@ -29,7 +29,7 @@ model =
                    putStrLn ""
                    
      runEventInStartTime $
-       enqueueEventWithIntegTimes $
+       enqueueEventWithLatticeTimes $
        showLatticeNode "enqueue"
 
      let reduce a b =
@@ -48,6 +48,6 @@ model =
 
 main :: IO ()
 main =
-  do lattice <- newRandomLattice 5
-     runLIO lattice $
+  do lat <- newRandomLattice 4
+     runLIO lat $
        runSimulation model specs