+++ tests/TraversingLattice4.hs
@@ -10,7 +10,7 @@ meanRepairTime = 0.5
 
 specs = Specs { spcStartTime = 0.0,
                 spcStopTime = 1000.0,
-                spcDT = 400.0,
+                spcDT = 0.1,
                 spcMethod = RungeKutta4,
                 spcGeneratorType = SimpleGenerator }
         
@@ -31,7 +31,7 @@ model =
      r <- newRef 0
                    
      runEventInStartTime $
-       enqueueEventWithIntegTimes $
+       enqueueEventWithLatticeTimes $
        do x <- liftParameter $ randomUniform 0 1
           writeRef r x
           showLatticeNode ("enqueue (x = " ++ show x ++ ")") 
@@ -59,6 +59,6 @@ model =
 
 main :: IO ()
 main =
-  do lattice <- newRandomLattice 5
-     runLIO lattice $
+  do lat <- newRandomLattice 4
+     runLIO lat $
        runSimulation model specs